Cleanup Driver, ExecutionEngine, TreadleTester API
This extends the API for interacting with things that need an
ExecutionOptionsManager to do their job. Specifically, this adds the
following two APIs in companion objects (if they did not already exist):
  - apply(optionsManager: HasTreadleSuite)
  - apply(args: Array[String])
The original API is then deprecated:
  - apply(firrtlString: String, optionsManager: HasTreadleSuite)

Tests are cleaned up to use (mostly) the Array[String] API due to its
brevity.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>