Quick fix (hack) for Stage/Phase CircuitForm issues. (#106)
Provide an argument to TreadleTester to specify what form the circuit is in,
Define PrepareAstFromLowFIRRTL object which avoids the "normal" FIRRTL transforms if we're already in LowForm. NOTE - trying to refactor this so the two sets of transforms (FIRRTL and Treadle) are composable doesn't appear to work.
Define checkFormTransform method in TreadleCompatibilityPhase which determines which phases are to be run based on the circuit form.